requirements.txt
functions-framework==3.*
requests
beautifulsoup4
pandas
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib